requirements.txt
asgiref==3.7.2
attrs==23.1.0
cffi==1.15.1
cryptography==40.0.2
Django==4.2.1
django-cors-headers==4.1.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
inflection==0.5.1
jsonschema==4.17.3
pycparser==2.21
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
typing_extensions==4.6.2
tzdata==2023.3
uritemplate==4.1.1
